feat: plant 정보, 종류 정보, 리스트 조회

+++ backend/src/main/java/com/plog/backend/domain/plant/dto/request/PlantAddRequest.java
@@ -11,8 +11,8 @@ import java.util.Date;
 @Builder
 @ToString
 public class PlantAddRequest {
-    int plantTypeId;
-    int otherPlantTypeId;
+    Long plantTypeId;
+    Long otherPlantTypeId;
     String profile;
     String bio;
     String nickname;
